pyproject.toml
[tool.poetry]
name = "homeworkproject"
version = "0.1.0"
description = ""
authors = ["Artyukhin Konstantin <k89161273297@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
openpyxl = "^3.1.4"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
mypy = "^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
warn_return_any = true
check_untyped_defs = true
disallow_untyped_defs = true
exclude = 'venv'


[tool.black]
line-length = 119
exclude = 'git'


[tool.isort]
line_length = 119
